fix(domain): テストコードを修正
